fix: rename app to frappe-private-comment
re: #17
