--- frappe_comment_xt/modules.txt
@@ -1,1 +1,0 @@
-Frappe Comment xt
