--- frappe_comment_xt/public/js/frappe_comment_xt.js
@@ -1,3 +1,0 @@
-frappe.require(["/assets/frappe_comment_xt/js/controls/timeline.js"]);
-frappe.require(["/assets/frappe_comment_xt/js/controls/footer.js"]);
-frappe.require(["/assets/frappe_comment_xt/js/controls/comment.js"]);
