+++ frappe_private_comment/__init__.py
@@ -5,7 +5,8 @@ __version__ = "0.0.1"
 
 def patch_add_comments_in_timeline():
     import frappe.desk.form.load as frappe_load
-    from frappe_comment_xt.helpers.comment import add_comments_in_timeline
+
+    from frappe_private_comment.helpers.comment import add_comments_in_timeline
 
     # A monkey patch was written for this function as it is used in many places within Frappe. Care was taken to avoid breaking existing code.
