+++ frappe_private_comment/frappe_private_comment/custom/comment.json
