+++ frappe_private_comment/modules.txt
@@ -1,0 +1,1 @@
+Frappe Private Comment
