+++ frappe_private_comment/public/js/frappe_private_comment.js
@@ -1,0 +1,3 @@
+frappe.require(["/assets/frappe_private_comment/js/controls/timeline.js"]);
+frappe.require(["/assets/frappe_private_comment/js/controls/footer.js"]);
+frappe.require(["/assets/frappe_private_comment/js/controls/comment.js"]);
